Rename components to be consistent
